refactor/feat: created shared Searchbar
