+++ src/Components/Shared/SearchBar/SearchBar.module.css
@@ -3,7 +3,8 @@
     align-items: center;
     border-radius: 4px;
     border: 0.5px solid #b2b2b2;
-    width: 30vw;
+    width: 20vw;
+    height: 40px;
     gap: 5px;
     justify-content: space-between;
 }
@@ -12,14 +13,11 @@
 
     font-family: Arial;
     width: inherit;
-    height: inherit;
     padding: 0.5rem;
     outline: none;
 }
 
 .searchBtn {
-    /* position: absolute;
-    right: 0; */
     height: max-content;
     background-color: transparent;
     cursor: pointer;
