+++ src/Components/workspace/Instances/Instances.module.css
@@ -1,10 +1,17 @@
 .container {
     display: grid;
-    grid-template-rows: repeat(3, 1fr);
+    grid-template-rows: auto auto;
 }
 
 .header {
-    width: fit-content;
+    grid-row: 1/2;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
+
+.title {
+    font-size: 1.5rem;
 }
 
 .cardsContainer {
