+++ src/Components/workspace/domains and hosting/DomainsAndHosting.js
@@ -1,15 +1,22 @@
-import SearchBar from './SearchBar';
-import { useSelector } from 'react-redux';
+import SearchBar from '../../Shared/SearchBar'
+import { useSelector, useDispatch } from 'react-redux';
+import { searchDomain } from '@/redux/actions/domains';
 import styles from './DomainsAndHosting.module.css';
 
 const Domains_and_hosting = () => {
   const { resultantDomains } = useSelector(state => state.domains);
+
+  const dispatch = useDispatch();
+
+  const handleSearch = (domain) => {
+    dispatch(searchDomain(domain));
+  }
   return (
     <div>
       <div>
         <h2>Customize your domain</h2>
         <h4>Check availability</h4>
-        <SearchBar/>
+        <SearchBar handleSearch={handleSearch} placeholder='Search a domain'/>
          {resultantDomains && resultantDomains.length > 0 ? (
         <table className={styles.table}>
           <thead>
