--- src/Components/workspace/domains and hosting/SearchBar/SearchBar.module.css
@@ -1,43 +1,0 @@
-.searchItems {
-    display: flex;
-    align-items: center;
-    border-radius: 4px;
-    border: 0.5px solid #b2b2b2;
-    width: 30vw;
-    gap: 5px;
-    justify-content: space-between;
-}
-
-.input {
-    
-    font-family: Arial;
-    width: inherit;
-    height: inherit;
-    padding: 0.5rem;
-    outline: none;
-}
-
-.searchBtn {
-    /* position: absolute;
-    right: 0; */
-    height: max-content;
-    background-color: transparent;
-    cursor: pointer;
-}
-
-.searchBtn:hover {
-    color: hsl(233, 9%, 79%);
-
-}
-
-.searchIcon {
-    background-color: rgba(128, 128, 128, 0.5);
-    transition: fill 0.3s ease;
-    border-radius: 50%;
-    padding: 3px;
-}
-
-.searchIcon:hover {
-    background-color: gray;
-    font-weight: bold;
-}
